+++ src/pages/About.vue
@@ -1,7 +1,16 @@
 <script setup lang="ts">
 import { useHead } from '@vueuse/head';
+import { config } from '@/config';
 
-useHead({ title: 'About - Tool.David888.com' });
+useHead({
+  title: 'About - Tool.David888.com',
+  link: [
+    {
+      rel: 'canonical',
+      href: `${config.app.siteOrigin}/about`,
+    },
+  ],
+});
 </script>
 
 <template>